Fixes #48 Seems like over the time yEd has changed how it reads the graphml file (but I can find no documentation exposing that)
