+++ pyyed/__init__.py
@@ -327,7 +327,7 @@ class Node:
         ET.SubElement(shape, "y:Shape", type=self.shape)
 
         if self.UML:
-            UML = ET.SubElement(shape, "y:UML")
+            UML = ET.SubElement(data, "y:UML")
 
             attributes = ET.SubElement(UML, "y:AttributeLabel", type=self.shape)
             attributes.text = self.UML["attributes"]
